A2048.Board

Description

Documentation

type Board = [[Int]]

Game board: board = [row], row = [tile].

type Game x = ReaderT Game2048Config (State Board) x

The 2048 game monad, a concrete instance for Monad2048 constraint.

type Monad2048 m = (MonadState Board m, MonadReader Game2048Config m)

All game actions run in a Monad2048 monad.

gridPos :: Game2048Config -> (Double, Double) -> (Double, Double)

Position of the selected grid, pure function.

translateGrid :: Game2048Config -> Double -> Double -> SVG -> SVG

Translate to the selected grid, pure function.

translateGridM :: MonadReader Game2048Config m => Double -> Double -> m SVG -> m SVG

Translate to the selected grid.

foreachGrid :: Monad2048 m => (Int -> m SVG) -> m [SVG]

Traverse all grids.

foreachNonEmptyGrid :: Monad2048 m => (Int -> m SVG) -> m [SVG]

Traverse all non-empty grids.

boardSVG :: Monad2048 m => m SVG

Generate an SVG image for the empty board.

snapshot :: Monad2048 m => m SVG

Take a snapshot of the current game status.

hold :: Monad2048 m => Double -> m Animation

Emit a static animation for the current game status.

holdWith :: Monad2048 m => TileLabelMode -> Double -> m Animation

Emit a static animation for the current game status.

gameAnimation :: Game2048Config -> Game a -> a

Convert a Monad2048 action to an animation.

mkPure :: Monad2048 m => (Time -> Game a) -> m (Time -> a)

Make a pure Animation from a Game monad.