A2048.Logic

Description

Documentation

data GameEvent a p

Game Events, parameterised by value and position.

Constructors

TileMerge a p p
This tile is merged from two tiles, original positions provided.

TileMove a p
This tile is merely moving, original position provided.

TileVanish
Placeholder, here no tile exists.

rowMerge :: (Eq a, Enum a) => [Maybe a] -> [GameEvent a Int]

Merge a row, Nothing as empty tiles.

boardMerge :: (Eq a, Enum a) => [[Maybe a]] -> [[GameEvent a (Int, Int)]]

Merge a whole board, Nothing as empty tiles.

data Direction

4 directions, for game moves.

Constructors

Left

Right

Up

Down

idxTrans :: Int -> Int -> Direction -> (Int, Int) -> (Int, Int)

Transform indices according to a Direction. This addIndices pairs each element with an index:

addIndices xs =
  [ [ (x, y, a)
    | (x, a) <- zip [0 ..] row]
  | (y, row) <- zip [0 ..] xs]

And this transIndices use idxTrans to transform indices:

transIndices = map (map (idxTrans w h d))

Then we have the following property:

rotateTo d . addIndices == transIndices . addIndices . rotateTo d

idxTrans transforms the index as if it were added before a rotateTo.

rotateTo :: Direction -> [[a]] -> [[a]]

Rotate game board according to a Direction.

rotateFrom :: Direction -> [[a]] -> [[a]]

Recover from a rotated game board according to a Direction.

events :: Int -> Int -> Direction -> [[Int]] -> [[GameEvent Int (Int, Int)]]

Gather game events on a game move to a given Direction.