A2048.Text

Description

Documentation

label :: Text -> SVG

Plain text labels, with Source Sans Pro, Source Han fonts.

colourLabel :: Texture -> Text -> SVG

Coloured text labels, with Source Sans Pro, Source Han fonts.